@@ -8,7 +8,7 @@
     <parent>
         <groupId>net.n2oapp.platform</groupId>
         <artifactId>n2o-boot-dependencies</artifactId>
-        <version>4.0.2-SNAPSHOT</version>
+        <version>4.2.7-SNAPSHOT</version>
         <relativePath>../n2o-platform-dependencies</relativePath>
     </parent>
 
@@ -20,7 +20,7 @@
         <maven.javadoc.version>3.2.0</maven.javadoc.version>
         <maven.javadoc.failOnError>false</maven.javadoc.failOnError>
 
-        <java.version>11</java.version>
+        <java.version>14</java.version>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -32,6 +32,14 @@
         <jacoco.aggregate.report.path>${project.basedir}/../target/site/jacoco-aggregate</jacoco.aggregate.report.path> <!-- By default use parent directory of module for multimodule projects. -->
 
         <config.server.enable>false</config.server.enable> <!-- Disable configuration server by default. To enable it use `consul` profile. -->
+
+        <sonar.sources>.</sonar.sources>
+        <sonar.inclusions>
+            src/main/resources/db/**,
+            src/main/resources/**/*.yml,
+            src/main/java/**
+        </sonar.inclusions>
+        <sonar.test.exclusions>src/test/**</sonar.test.exclusions>
     </properties>
 
     <dependencies>
@@ -81,7 +89,15 @@
                         <execution>
                             <goals>
                                 <goal>repackage</goal>
+                                <goal>build-info</goal>
                             </goals>
+                            <configuration>
+                                <additionalProperties>
+                                    <java.source>${maven.compiler.source}</java.source>
+                                    <java.target>${maven.compiler.target}</java.target>
+                                    <config.server.enable>${config.server.enable}</config.server.enable>
+                                </additionalProperties>
+                            </configuration>
                         </execution>
                     </executions>
                     <configuration>
