@@ -1,18 +1,23 @@
 package net.n2oapp.platform.jaxrs;
 
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
 import java.io.Serializable;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * Сообщение, возвращаемое в ответе REST сервиса
  */
 public class RestMessage implements Serializable {
+
     private static final long serialVersionUID = 139886274946702785L;
     private String message;
-    private List<Error> errors;
+    private List<? extends BaseError> errors;
     private String[] stackTrace;
 
-    public RestMessage(List<Error> errors) {
+    public RestMessage(List<? extends BaseError> errors) {
         this.errors = errors;
     }
 
@@ -23,45 +28,11 @@ public class RestMessage implements Serializable {
     public RestMessage() {
     }
 
-    public static class Error implements Serializable {
-        private static final long serialVersionUID = -6241752723478340674L;
-        private String field;
-        private String message;
-
-        public Error() {
-        }
-
-        public Error(String field, String message) {
-            this.field = field;
-            this.message = message;
-        }
-
-        public Error(String message) {
-            this(null, message);
-        }
-
-        public String getField() {
-            return field;
-        }
-
-        public void setField(String field) {
-            this.field = field;
-        }
-
-        public String getMessage() {
-            return message;
-        }
-
-        public void setMessage(String message) {
-            this.message = message;
-        }
-    }
-
     public String getMessage() {
         return message;
     }
 
-    public List<Error> getErrors() {
+    public List<? extends BaseError> getErrors() {
         return errors;
     }
 
@@ -72,4 +43,128 @@ public class RestMessage implements Serializable {
     public void setStackTrace(String[] stackTrace) {
         this.stackTrace = stackTrace;
     }
+
+    @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "t")
+    @JsonSubTypes({
+            @JsonSubTypes.Type(value = Error.class, name = "1"),
+            @JsonSubTypes.Type(value = ConstraintViolationError.class, name = "2")
+    })
+    public abstract static class BaseError implements Serializable {
+
+        private static final long serialVersionUID = -6241752723478340674L;
+
+        private String message;
+
+        public BaseError() {
+        }
+
+        public BaseError(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+
+        public void setMessage(String message) {
+            this.message = message;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+            BaseError error = (BaseError) o;
+            return Objects.equals(getMessage(), error.getMessage());
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(getMessage());
+        }
+
+    }
+
+    public static class Error extends BaseError {
+
+        public Error() {
+        }
+
+        public Error(String message) {
+            super(message);
+        }
+
+    }
+
+    public static class ConstraintViolationError extends BaseError {
+
+        private static final long serialVersionUID = -1337L;
+
+        private String field;
+        private String leafBeanClass;
+        private String rootBeanClass;
+        private String constraintAnnotation;
+
+        public ConstraintViolationError() {
+        }
+
+        public ConstraintViolationError(String message, String field, String leafBeanClass, String rootBeanClass, String constraintAnnotation) {
+            super(message);
+            this.field = field;
+            this.leafBeanClass = leafBeanClass;
+            this.rootBeanClass = rootBeanClass;
+            this.constraintAnnotation = constraintAnnotation;
+        }
+
+        public String getField() {
+            return field;
+        }
+
+        public void setField(String field) {
+            this.field = field;
+        }
+
+        public String getLeafBeanClass() {
+            return leafBeanClass;
+        }
+
+        public void setLeafBeanClass(String leafBeanClass) {
+            this.leafBeanClass = leafBeanClass;
+        }
+
+        public String getRootBeanClass() {
+            return rootBeanClass;
+        }
+
+        public void setRootBeanClass(String rootBeanClass) {
+            this.rootBeanClass = rootBeanClass;
+        }
+
+        public String getConstraintAnnotation() {
+            return constraintAnnotation;
+        }
+
+        public void setConstraintAnnotation(String constraintAnnotation) {
+            this.constraintAnnotation = constraintAnnotation;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+            if (!super.equals(o)) return false;
+            ConstraintViolationError that = (ConstraintViolationError) o;
+            return Objects.equals(getField(), that.getField()) &&
+                    Objects.equals(getLeafBeanClass(), that.getLeafBeanClass()) &&
+                    Objects.equals(getRootBeanClass(), that.getRootBeanClass()) &&
+                    Objects.equals(getConstraintAnnotation(), that.getConstraintAnnotation());
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(super.hashCode(), getField(), getLeafBeanClass(), getRootBeanClass(), getConstraintAnnotation());
+        }
+
+    }
+
 }
