@@ -6,6 +6,7 @@ import org.apache.cxf.phase.AbstractPhaseInterceptor;
 import org.apache.cxf.phase.Phase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.OAuth2ClientContext;
+import org.springframework.security.oauth2.client.OAuth2RestTemplate;
 import org.springframework.security.oauth2.common.OAuth2AccessToken;
 import org.springframework.util.StringUtils;
 
@@ -21,7 +22,7 @@ import java.util.Map;
 public class JaxRsJwtHeaderInterceptor extends AbstractPhaseInterceptor<Message> {
 
     @Autowired(required = false)
-    private OAuth2ClientContext oauth2ClientContext;
+    private OAuth2RestTemplate restTemplate;
 
     public JaxRsJwtHeaderInterceptor() {
         super(Phase.WRITE);
@@ -30,15 +31,17 @@ public class JaxRsJwtHeaderInterceptor extends AbstractPhaseInterceptor<Message>
     @SuppressWarnings("unchecked")
     @Override
     public void handleMessage(Message message) {
-        if (oauth2ClientContext != null
-                && oauth2ClientContext.getAccessToken() != null) {
-            String tokenType = oauth2ClientContext.getAccessToken().getTokenType();
-            if (!StringUtils.hasText(tokenType)) {
-                tokenType = OAuth2AccessToken.BEARER_TYPE;
+        if (restTemplate != null) {
+            OAuth2AccessToken accessToken = restTemplate.getAccessToken();
+            if (accessToken != null) {
+                String tokenType = accessToken.getTokenType();
+                if (!StringUtils.hasText(tokenType)) {
+                    tokenType = OAuth2AccessToken.BEARER_TYPE;
+                }
+                Map<String, List> headers = (Map<String, List>) message.get("org.apache.cxf.message.Message.PROTOCOL_HEADERS");
+                headers.put("Authorization", Collections.singletonList(String.format("%s %s", tokenType,
+                        accessToken.getValue())));
             }
-            Map<String, List> headers = (Map<String, List>) message.get("org.apache.cxf.message.Message.PROTOCOL_HEADERS");
-            headers.put("Authorization", Collections.singletonList(String.format("%s %s", tokenType,
-                    oauth2ClientContext.getAccessToken().getValue())));
         }
     }
 }
