@@ -17,6 +17,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.security.oauth2.client.OAuth2ClientContext;
 import org.springframework.security.oauth2.client.OAuth2RestTemplate;
 import org.springframework.security.oauth2.client.resource.OAuth2ProtectedResourceDetails;
@@ -26,6 +27,7 @@ import java.text.SimpleDateFormat;
 @Configuration
 @AutoConfigureAfter(name = "org.springframework.boot.autoconfigure.security.oauth2.OAuth2AutoConfiguration")
 @AutoConfigureBefore(N2oFrameworkAutoConfiguration.class)
+@PropertySource("classpath:web.n2o.default.properties")
 public class WebAutoConfiguration {
 
     @ConditionalOnMissingBean
@@ -47,6 +49,9 @@ public class WebAutoConfiguration {
         @Value("${n2o.engine.rest.dateformat.deserialize}")
         private String[] deserializingFormats;
 
+        @Value("${n2o.engine.rest.dateformat.exclusion-keys}")
+        private String[] exclusionKeys;
+
         @Bean("oauth2RestTemplate")
         public OAuth2RestTemplate oauth2RestTemplate(OAuth2ProtectedResourceDetails details, OAuth2ClientContext oauth2ClientContext) {
             OAuth2RestTemplate restTemplate = new OAuth2RestTemplate(details, oauth2ClientContext);
@@ -60,7 +65,7 @@ public class WebAutoConfiguration {
                                                                         @Value("${n2o.engine.rest.url}") String baseRestUrl) {
             ObjectMapper restObjectMapper = new ObjectMapper();
             restObjectMapper.setDateFormat(new SimpleDateFormat(serializingFormat));
-            RestEngineTimeModule module = new RestEngineTimeModule(deserializingFormats);
+            RestEngineTimeModule module = new RestEngineTimeModule(deserializingFormats, exclusionKeys);
             restObjectMapper.registerModules(module);
             SpringRestDataProviderEngine springRestDataProviderEngine = new SpringRestDataProviderEngine(oauth2RestTemplate, restObjectMapper);
             springRestDataProviderEngine.setBaseRestUrl(baseRestUrl);
