@@ -14,7 +14,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.security.oauth2.client.OAuth2ClientContext;
+import org.springframework.security.oauth2.client.token.DefaultAccessTokenRequest;
 import org.springframework.security.oauth2.common.DefaultOAuth2AccessToken;
+import org.springframework.security.oauth2.common.DefaultOAuth2RefreshToken;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.web.client.RestTemplate;
@@ -24,6 +26,7 @@ import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
@@ -37,15 +40,17 @@ import static org.hamcrest.core.Is.is;
         properties = {
                 "server.port=62517",
                 "security.oauth2.client.client-id=test",
+                "security.oauth2.client.client-secret=test",
+                "security.oauth2.client.access-token-uri=http://localhost:${server.port}/token",
                 "cxf.jaxrs.client.classes-scan=true",
                 "cxf.jaxrs.client.classes-scan-packages=net.n2oapp.platform.web",
-                "cxf.jaxrs.client.address=http://localhost:${server.port}/token"
+                "cxf.jaxrs.client.address=http://localhost:${server.port}/protectedResource"
         })
 @SpringBootApplication
 public class JwtForwardingTest {
 
-    private static final String jwt = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6Ik" +
-            "pvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c";
+    public static final String accessToken = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ";
+    public static final String refreshedToken = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyLCJyZWZyZXNoZWQiOnRydWV9";
 
     private static final RestTemplate restTemplate = new RestTemplate();
 
@@ -60,23 +65,36 @@ public class JwtForwardingTest {
 
     @Before
     public void setUp() {
-        DefaultOAuth2AccessToken token = new DefaultOAuth2AccessToken(jwt);
+        DefaultOAuth2AccessToken token = new DefaultOAuth2AccessToken(accessToken);
         token.setTokenType("Bearer");
         Mockito.when(clientContext.getAccessToken()).thenReturn(token);
+        Mockito.when(clientContext.getAccessTokenRequest()).thenReturn(new DefaultAccessTokenRequest());
     }
 
     @Test
     public void testRestProvider() {
         builder.routes(new RouteInfo("/testRestTemplate", new QueryContext("testRestTemplate")));
         Map result = restTemplate.getForObject("http://localhost:" + port + "/n2o/data/testRestTemplate", Map.class);
-        assertThat(((Map) ((List) result.get("list")).get(0)).get("auth"), is("Bearer " + jwt));
+        assertThat(((Map) ((List) result.get("list")).get(0)).get("auth"), is("Bearer " + accessToken));
     }
 
     @Test
     public void testJaxRsProxyClient() {
         builder.routes(new RouteInfo("/testJaxRs", new QueryContext("testJaxRs")));
         Map result = restTemplate.getForObject("http://localhost:" + port + "/n2o/data/testJaxRs", Map.class);
-        assertThat(((Map) ((List) result.get("list")).get(0)).get("auth"), is("Bearer " + jwt));
+        assertThat(((Map) ((List) result.get("list")).get(0)).get("auth"), is("Bearer " + accessToken));
+    }
+
+    @Test
+    public void testJaxRsProxyClientRefreshToken() {
+        DefaultOAuth2AccessToken token = new DefaultOAuth2AccessToken(accessToken);
+        token.setRefreshToken(new DefaultOAuth2RefreshToken(accessToken));
+        token.setExpiration(new Date(0));
+        Mockito.when(clientContext.getAccessToken()).thenReturn(token);
+
+        builder.routes(new RouteInfo("/testJaxRs", new QueryContext("testJaxRs")));
+        Map result = restTemplate.getForObject("http://localhost:" + port + "/n2o/data/testJaxRs", Map.class);
+        assertThat(((Map) ((List) result.get("list")).get(0)).get("auth"), is("Bearer " + refreshedToken));
     }
 
 
