@@ -2,6 +2,8 @@ package net.n2oapp.platform.web.test;
 
 import net.n2oapp.framework.api.exception.N2oException;
 import net.n2oapp.framework.api.exception.N2oUserException;
+import net.n2oapp.framework.api.exception.ValidationMessage;
+import net.n2oapp.framework.api.metadata.global.dao.object.N2oObject;
 import net.n2oapp.framework.api.metadata.local.CompiledObject;
 import net.n2oapp.platform.i18n.UserException;
 import net.n2oapp.platform.jaxrs.RestException;
@@ -14,11 +16,15 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.client.HttpClientErrorException;
 
 import java.nio.charset.Charset;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Stream;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.instanceOf;
 
 /**
  * Тесты класса {@link PlatformExceptionHandler}
@@ -109,5 +115,40 @@ public class PlatformExceptionHandlerTest {
         assertThat(e.getMessage(), containsString("Unexpected error"));
         assertThat(Stream.of(ExceptionUtils.getStackFrames(e.getCause())).anyMatch(sf ->
                 sf.contains("something went wrong")), is(true));
+    }
+
+    /**
+     * Тест обработки пользовательских сообщений из множественных ошибок, выбрасываемых JaxRs Proxy Client
+     */
+    @Test
+    public void handleMultipleErrorsFromJaxRsClient() {
+        PlatformExceptionHandler handler = new PlatformExceptionHandler();
+
+        N2oObject.Parameter inParam1 = new N2oObject.Parameter();
+        inParam1.setValidationFailKey("create.arg0.name");
+        N2oObject.Parameter inParam2 = new N2oObject.Parameter();
+        inParam2.setValidationFailKey("create.arg0.age");
+        CompiledObject.Operation operation = new CompiledObject.Operation(
+                Map.of("name", inParam1, "age", inParam2), null);
+
+        RestMessage restMessage = new RestMessage(Arrays.asList(
+                new RestMessage.ConstraintViolationError("Не должно равняться нулю", "create.arg0.name", null, null, null),
+                new RestMessage.Error("Wrong data"),
+                new RestMessage.Error("Wrong format"),
+                new RestMessage.ConstraintViolationError("Не должно быть отрицательным", "create.arg0.age", null, null, null)
+        ));
+
+        N2oException e = handler.handle(operation, null,
+                new N2oException(new RestException(restMessage, 400)));
+        assertThat(e, instanceOf(N2oUserException.class));
+        N2oUserException userException = (N2oUserException) e;
+        assertThat(userException.getUserMessage(), is("1) Wrong data\n2) Wrong format\n"));
+
+        List<ValidationMessage> messages = userException.getMessages();
+        assertThat(messages.size(), is(2));
+        assertThat(messages.get(0).getMessage(), is("Не должно равняться нулю"));
+        assertThat(messages.get(0).getFieldId(), is("name"));
+        assertThat(messages.get(1).getMessage(), is("Не должно быть отрицательным"));
+        assertThat(messages.get(1).getFieldId(), is("age"));
     }
 }
