@@ -2,26 +2,40 @@ package net.n2oapp.platform.web.test;
 
 import net.n2oapp.criteria.dataset.DataSet;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.authentication.TestingAuthenticationToken;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.common.DefaultOAuth2AccessToken;
+import org.springframework.security.oauth2.common.DefaultOAuth2RefreshToken;
+import org.springframework.security.oauth2.common.OAuth2AccessToken;
+import org.springframework.security.web.authentication.AnonymousAuthenticationFilter;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.core.MediaType;
+import java.io.IOException;
 import java.util.Collections;
+import java.util.Date;
 import java.util.Map;
 
 @Configuration
-@RestController
 public class TestConfiguration extends WebSecurityConfigurerAdapter {
+
     @Override
     protected void configure(HttpSecurity http) throws Exception {
-        http.authorizeRequests().anyRequest().permitAll();
-    }
-
-    @GetMapping(value = "/token", produces = MediaType.APPLICATION_JSON)
-    public DataSet token(HttpServletRequest request) {
-        return new DataSet("token", Collections.singletonList(Map.of("auth", request.getHeader("authorization"))));
+        http.authorizeRequests().anyRequest().permitAll().and().addFilterBefore(new Filter() {
+            @Override
+            public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+                TestingAuthenticationToken authenticationToken = new TestingAuthenticationToken("user", "N/A");
+                authenticationToken.setAuthenticated(true);
+                SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+                filterChain.doFilter(servletRequest, servletResponse);
+            }
+        }, AnonymousAuthenticationFilter.class).csrf().disable();
     }
 }
